DXP-1451 Merge product attribute provider classes
